requirements.txt
fastapi>=0.109.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
httpx>=0.26.0
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=2.3.0
pgvector>=0.2.4
apscheduler>=3.10.4
pydantic>=2.5.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
feedparser>=6.0.10
uvicorn>=0.27.0
jinja2>=3.1.2
python-multipart>=0.0.6
